fix: replace mock and rename button
