+++ src/app/chart/_components/FetchDataButton.tsx
@@ -1,11 +1,11 @@
 'use client'
 import { FitnessOutput } from '@/common/LLM/fitnessOutputSchema'
-import { fetchRecommendedData } from '@/utils/fetchAPIData'
+import { fetchRecommendedDataWithMock } from '@/utils/fetchAPIData'
 
 export default function FetchDataButton({ fitnessData }: { fitnessData: FitnessOutput }) {
   const handleClick = async () => {
     try {
-      const recommendedData = await fetchRecommendedData(fitnessData, null)
+      const recommendedData = await fetchRecommendedDataWithMock(fitnessData, null)
       console.log('Recommended Data:', recommendedData)
     } catch (error) {
       console.error('Error fetching recommended data:', error)
